Clean up utils and remove duplicate modules
