+++ app/utils/security.py
@@ -1,0 +1,38 @@
+from passlib.context import CryptContext
+from jose import JWTError, jwt
+from datetime import datetime, timedelta
+from typing import Optional
+import os
+
+pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
+
+def verify_password(plain_password: str, hashed_password: str) -> bool:
+    return pwd_context.verify(plain_password, hashed_password)
+
+def get_password_hash(password: str) -> str:
+    return pwd_context.hash(password)
+
+def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.utcnow() + expires_delta
+    else:
+        expire = datetime.utcnow() + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(
+        to_encode,
+        os.getenv("JWT_SECRET_KEY", "CHANGEME"),
+        algorithm=os.getenv("JWT_ALGORITHM", "HS256")
+    )
+    return encoded_jwt
+
+def verify_token(token: str) -> Optional[dict]:
+    try:
+        payload = jwt.decode(
+            token,
+            os.getenv("JWT_SECRET_KEY", "CHANGEME"),
+            algorithms=[os.getenv("JWT_ALGORITHM", "HS256")]
+        )
+        return payload
+    except JWTError:
+        return None
